LEARNER-4146, fixed comment, code improvements
