+++ Source/CutomePlayer/PlayerView.swift
@@ -1,0 +1,30 @@
+//
+//  PlayerView.swift
+//  edX
+//
+//  Created by Salman on 09/04/2018.
+//  Copyright © 2018 edX. All rights reserved.
+//
+
+import UIKit
+import AVKit
+
+class PlayerView: UIView {
+    var player: AVPlayer? {
+        get {
+            return playerLayer.player
+        }
+        
+        set {
+            playerLayer.player = newValue
+        }
+    }
+    
+    var playerLayer: AVPlayerLayer {
+        return layer as! AVPlayerLayer
+    }
+    
+    override class var layerClass: AnyClass {
+        return AVPlayerLayer.self
+    }
+}
